+++ src/docker.rs
@@ -1,8 +1,9 @@
-use std::{env, fs};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::{Command, ExitStatus};
+use std::{env, fs};
 
-use Target;
+use {Target, Toml};
+use cargo::Root;
 use errors::*;
 use extensions::CommandExt;
 use id;
@@ -23,12 +24,14 @@ pub fn register(verbose: bool) -> Result<()> {
         .run(verbose)
 }
 
-pub fn run(target: Target,
-           docker_image: String,
+pub fn run(target: &Target,
            args: &[String],
-           cargo_root: &Path,
+           root: &Root,
+           toml: Option<&Toml>,
+           uses_xargo: bool,
            verbose: bool)
            -> Result<ExitStatus> {
+    let root = root.path();
     let home_dir = env::home_dir()
         .ok_or_else(|| "couldn't get home directory. Is $HOME not set?")?;
     let cargo_dir = env::var_os("CARGO_HOME")
@@ -37,7 +40,7 @@ pub fn run(target: Target,
     let xargo_dir = env::var_os("XARGO_HOME")
         .map(PathBuf::from)
         .unwrap_or_else(|| home_dir.join(".xargo"));
-    let target_dir = cargo_root.join("target");
+    let target_dir = root.join("target");
 
     // create the directories we are going to mount before we mount them,
     // otherwise `docker` will create them but they will be owned by `root`
@@ -45,16 +48,16 @@ pub fn run(target: Target,
     fs::create_dir(&cargo_dir).ok();
     fs::create_dir(&xargo_dir).ok();
 
-    let mut cmd = if target.uses_xargo() {
+    let mut cmd = if uses_xargo {
         Command::new("xargo")
     } else {
         Command::new("cargo")
     };
     cmd.args(args);
 
-    let cargo_lock = cargo_root.join("Cargo.lock");
+    let cargo_lock = root.join("Cargo.lock");
     if !cargo_lock.exists() {
-        let cargo_toml = cargo_root.join("Cargo.toml");
+        let cargo_toml = root.join("Cargo.toml");
         Command::new("cargo").args(&["generate-lockfile",
                     "--manifest-path",
                     &cargo_toml.display().to_string()])
@@ -72,12 +75,29 @@ pub fn run(target: Target,
         .args(&["-e", "XARGO_HOME=/xargo"])
         .args(&["-v", &format!("{}:/xargo", xargo_dir.display())])
         .args(&["-v", &format!("{}:/cargo", cargo_dir.display())])
-        .args(&["-v", &format!("{}:/project:ro", cargo_root.display())])
+        .args(&["-v", &format!("{}:/project:ro", root.display())])
         .args(&["-v",
                 &format!("{}:/rust:ro", rustc::sysroot(verbose)?.display())])
         .args(&["-v", &format!("{}:/target", target_dir.display())])
         .args(&["-w", "/project"])
-        .args(&["-it", &docker_image])
+        .args(&["-it", &image(toml, target)?])
         .args(&["sh", "-c", &format!("PATH=$PATH:/rust/bin {:?}", cmd)])
         .run_and_get_status(verbose)
+}
+
+fn image(toml: Option<&Toml>, target: &Target) -> Result<String> {
+    Ok(if let Some(toml) = toml {
+            toml.image(target)?.map(|s| s.to_owned())
+        } else {
+            None
+        }
+        .unwrap_or_else(|| {
+            let version = env!("CARGO_PKG_VERSION");
+            let tag = if version.ends_with("-dev") {
+                "latest"
+            } else {
+                version
+            };
+            format!("japaric/{}:{}", target.triple(), tag)
+        }))
 }